order detail : involved_materials and involved_stations

+++ app/models.py
@@ -345,7 +345,7 @@ class Order(models.Model):
     product = models.ForeignKey(Product, on_delete=models.CASCADE,verbose_name = " محصول ")
     code = models.CharField(max_length=50,null=True, blank=True,verbose_name = "کد ")
     description = models.TextField(max_length=900,null=True, blank=True,verbose_name = "توضیحات")
-    circulation = models.IntegerField(default='1',verbose_name = " تیراژ ")
+    circulation = models.DecimalField(default='1',max_digits=30, decimal_places=4, verbose_name = " تیراژ ")
     confirmed = models.BooleanField(default=False, verbose_name = " تایید شده " )
     completed = models.BooleanField(default=False, verbose_name = " تکمیل شده " )
